requirements.txt
streamlit==1.32.0
yt-dlp==2024.3.10
openai-whisper==20240930
fugashi==1.2.1
unidic-lite==1.0.8
googletrans==4.0.0-rc1
gtts==2.5.4
genanki==0.13.1
pandas==2.0.2
torch
numpy
